lib: Move the library component to its own module.
This commit provides structure to the library component such that it
makes use of the new component framework.  By doing so we provide a
simpler way of interacting with the source code.
